src/2021/challenges/13/app.test.ts
import { expect, test } from 'vitest'
import { wrapGifts } from './app'

test(`Should return ['******',
'*🏈🎸*',
'*🎮🧸*',
'******'
]`, () => {
  const result = wrapGifts(['🏈🎸', '🎮🧸'])
  expect(result).toEqual(['******',
    '*🏈🎸*',
    '*🎮🧸*',
    '******'
  ])
})

test(`Should return [ '****',
'*📷*',
'*⚽️*',
'****'
]`, () => {
  const result = wrapGifts(['📷', '⚽️'])
  expect(result).toEqual(['****',
    '*📷*',
    '*⚽️*',
    '****'
  ])
})

test(`Should return [ '****',
'*📷*',
'****'
]`, () => {
  const result = wrapGifts(['📷'])
  expect(result).toEqual(['****',
    '*📷*',
    '****'
  ])
})

test('Should return []', () => {
  const result = wrapGifts([])
  expect(result).toEqual([])
})
